+++ src/Request/Model/ListCompanyAccounts.php
@@ -4,15 +4,18 @@ namespace Easir\SDK\Request\Model;
 
 use Easir\SDK\Request\Model;
 
-/**
- * The request model for ListCompanyAccounts
- *
- * @package Easir\SDK\Request\Model
- */
 class ListCompanyAccounts extends Model
 {
     /**
      * @var string
      */
     public $searchTerm = '';
+    /**
+     * @var integer
+     */
+    public $page = 1;
+    /**
+     * @var integer
+     */
+    public $perPage = 15;
 }