+++ src/Request/Model/ListCompanyUsers.php
@@ -4,11 +4,6 @@ namespace Easir\SDK\Request\Model;
 
 use Easir\SDK\Request\Model;
 
-/**
- * The request model for ListCompanyUsers
- *
- * @package Easir\SDK\Request\Model
- */
 class ListCompanyUsers extends Model
 {
     /**