+++ src/Response/ListCompanyTeams.php
@@ -6,10 +6,6 @@ use Easir\SDK\Model\Pagination;
 use Easir\SDK\Model\Team;
 use Easir\SDK\Response;
 
-/**
- * Class ListCompanyTeams
- * @package Easir\SDK\Response
- */
 class ListCompanyTeams extends Response
 {
     /**